bot/requirements.txt
aiohttp==3.8.3
discord.py==2.1.0
dateparser==1.1.6
mystbin.py==5.1.0 # this caused errors for me before, if it does for you, just remove the ==5.1.0 part
python-dotenv==0.21.0
better-ipc==2.0.1
humanize==4.4.0
psutil==5.9.4
Pillow==9.4.0
pilmoji==2.0.2
requests==2.28.2
expr.py==0.3.0
playwright==1.29.1
jishaku==2.5.1
beautifulsoup4==4.11.1